config: move cors config to separate file
